+++ R/tryCatchLog.R
@@ -1,4 +1,6 @@
 # ***************************************************************************
+# Copyright (C) 2016 Jürgen Altfeld (R@altfeld-im.de)
+# ---------------------------------------------------------------------------
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
 # the Free Software Foundation, either version 3 of the License, or
@@ -13,7 +15,7 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 # ***************************************************************************
 
-# source("R/dump.frames.dev.R")
+
 
 # # Do not use this since it makes all function of these packages available
 # # (even those who are not used by this package!)