+++ NEWS.md
@@ -30,3 +30,7 @@ This file describes the major changes of bug fixes in the package "tryCatchLog"
 
 * Added: Parameter `silent.messages` to `tryCatchLog` and `tryLog`
 * License: Added the copyright header to each R file to clarify the legal side
+
+## current version (ready-to-be-released features)
+
+* Improved: Documentation